update array dynamic test
