--- test/cases_source_filter_type_array_dynamic/1.btx_log.txt
@@ -1,1 +1,0 @@
-e_array_dynamic: { length = 3, dummy = 12, entries = [ [0] = 1, [1] = 2, [2] = 3 ] }
