--- test/cases_source_filter_type_array_dynamic/1.callbacks.c
@@ -1,12 +1,0 @@
-#include <metababel/metababel.h>
-
-void btx_push_usr_messages(void *btx_handle, void *usr_data,
-                           btx_source_status_t *status) {
-  long int foo[] = {1, 2, 3};
-  btx_push_message_e_array_dynamic(btx_handle, 3, 12, foo);
-  *status = BTX_SOURCE_END;
-}
-
-void btx_register_usr_callbacks(void *btx_handle) {
-  btx_register_callbacks_push_usr_messages(btx_handle, &btx_push_usr_messages);
-}
