+++ test/data_types/cases_array_dynamic/1.btx_log.txt
@@ -1,0 +1,1 @@
+event: { length = 3, entries = [ [0] = 1, [1] = 2, [2] = 3 ] }
