+++ test/data_types/test_array_dynamic.rb
@@ -1,0 +1,25 @@
+require 'base_test'
+
+class TestArrayDynamic < Test::Unit::TestCase
+  include GenericTest
+  extend VariableAccessor
+  include VariableClassAccessor
+
+  def self.startup
+    @btx_components = [
+      {
+        btx_component_type: 'SOURCE',
+        btx_component_downstream_model: './test/data_types/cases_array_dynamic/1.btx_model.yaml',
+        btx_file_usr_callbacks: './test/data_types/cases_array_dynamic/1.source_callbacks.c'
+      },
+      {
+        btx_component_type: 'FILTER',
+        btx_component_upstream_model: './test/data_types/cases_array_dynamic/1.btx_model.yaml',
+        btx_component_downstream_model: './test/data_types/cases_array_dynamic/1.btx_model.yaml',
+        btx_file_usr_callbacks: './test/data_types/cases_array_dynamic/1.filter_callbacks.c'
+      }
+    ]
+
+    @btx_output_validation = './test/data_types/cases_array_dynamic/1.btx_log.txt'
+  end
+end
